+++ qiskit/ignis/verification/tomography/basis/__init__.py
@@ -12,15 +12,25 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""
-Quantum tomography basis
+"""Quantum tomography basis
 """
 
 # Tomography Circuit Generation Functions
 from .tomographybasis import TomographyBasis
+from .gatesetbasis import GateSetBasis
+from .gatesetbasis import default_gateset_basis
 from .paulibasis import PauliBasis
 from .sicbasis import SICBasis
 from .circuits import state_tomography_circuits
 from .circuits import process_tomography_circuits
+from .circuits import gateset_tomography_circuits
 from .circuits import default_basis
 from .circuits import tomography_circuit_tuples
+
+from .paulibasis import pauli_measurement_circuit
+from .paulibasis import pauli_preparation_circuit
+from .paulibasis import pauli_measurement_matrix
+from .paulibasis import pauli_preparation_matrix
+
+from .sicbasis import sicpovm_preparation_circuit
+from .sicbasis import sicpovm_preparation_matrix